Delete most API server unit tests
