--- api/data_explorer/test/base_test_case.py
@@ -1,15 +1,0 @@
-import connexion
-import logging
-
-from flask_testing import TestCase
-
-from data_explorer.encoder import JSONEncoder
-
-
-class BaseTestCase(TestCase):
-    def create_app(self):
-        logging.getLogger('connexion.operation').setLevel('ERROR')
-        app = connexion.App(__name__, specification_dir='../swagger/')
-        app.app.json_encoder = JSONEncoder
-        app.add_api('swagger.yaml')
-        return app.app
