--- api/data_explorer/test/mock_responses/api_server_facets_basic.txt
@@ -1,28 +1,0 @@
-{
-   "count":1338,
-   "facets":[
-      {
-         "name":"Region",
-         "description":"Region description",
-         "es_field_name":"Region es field name",
-         "values":[
-            {
-               "count":364,
-               "name":"southeast"
-            },
-            {
-               "count":325,
-               "name":"northwest"
-            },
-            {
-               "count":325,
-               "name":"southwest"
-            },
-            {
-               "count":324,
-               "name":"northeast"
-            }
-         ]
-      }
-   ]
-}
