--- api/data_explorer/test/mock_responses/api_server_facets_histogram.txt
@@ -1,35 +1,0 @@
-{
-  "count": 1338,
-  "facets": [
-    {
-      "values": [
-        {
-          "count": 137,
-          "name": "10-20"
-        },
-        {
-          "count": 280,
-          "name": "20-30"
-        },
-        {
-          "count": 257,
-          "name": "30-40"
-        },
-        {
-          "count": 279,
-          "name": "40-50"
-        },
-        {
-          "count": 271,
-          "name": "50-60"
-        },
-        {
-          "count": 114,
-          "name": "60-70"
-        }
-      ],
-      "name": "Age",
-      "es_field_name": "Age es field name"
-    }
-  ]
-}
