--- api/data_explorer/test/mock_responses/es_histogram.txt
@@ -1,48 +1,0 @@
-{
-  "hits": {
-    "hits": [],
-    "total": 1338,
-    "max_score": 0.0
-  },
-  "timed_out": false,
-  "_shards": {
-    "successful": 5,
-    "failed": 0,
-    "skipped": 0,
-    "total": 5
-  },
-  "took": 5,
-  "aggregations": {
-    "_filter_Age es field name": {
-      "Age es field name": {
-        "buckets": [
-          {
-            "key": 10.0,
-            "doc_count": 137
-          },
-          {
-            "key": 20.0,
-            "doc_count": 280
-          },
-          {
-            "key": 30.0,
-            "doc_count": 257
-          },
-          {
-            "key": 40.0,
-            "doc_count": 279
-          },
-          {
-            "key": 50.0,
-            "doc_count": 271
-          },
-          {
-            "key": 60.0,
-            "doc_count": 114
-          }
-        ]
-      },
-      "doc_count": 1338
-    }
-  }
-}
